Component Reorganization
Created new subdirectories for components to make it easier to comprehend
